I added lava and a jump delay system
I added a jump delay thing, where if you click the jump button 6 frames or lower before hitting the ground, you will jump as soon as you land, this is used in most platformers. I also added a lava section, that will kill you once I figure out how to do hitboxes like that. It is also animated, which was a huge pain because I have never animated a tileset before, and because of my art style I also needed to animate all the ground tiles directly next to the lava.